index.ts,create-secret,render-qr-codes: move duplicate prompt function to common utils
See https://github.com/nimiq/cashlink-generator/pull/1#discussion_r1932852932
